pyproject.toml
[build-system]
requires = [
    "setuptools>=69",
    "wheel",
    "setuptools_scm>=8.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "activitysim"
dynamic = ["version"]
dependencies = [
    "cytoolz >= 0.8.1",
    "numba >= 0.57",
    "numpy >= 1.16.1, <2",
    "openmatrix >= 0.3.4.1",
    "pandas >= 1.4, <2",
    "pandera >=0.15, <0.18.1",
    "platformdirs",
    "psutil >= 4.1",
    "pyarrow >= 2.0",
    "pydantic >= 2.6",
    "pypyr >= 5.3",
    "pyyaml >= 5.1",
    "requests >= 2.7",
    "scikit-learn >= 1.1",
    "sharrow >= 2.9.1",
    "simwrapper > 1.7",
    "sparse",
    "tables >= 3.9",
    "xarray >= 0.21"
]
readme = "README.md"
requires-python = ">=3.10"

[project.scripts]
activitysim = "activitysim.cli.main:main"
activitysim_checksums = "activitysim.cli.create:display_sha256_checksums"

[tool.setuptools.packages.find]
exclude = [
    "other*", "test*",
    "activitysim/estimation/test*",
    "activitysim/abm/test*",
    "activitysim/cli/test*",
    "activitysim/core/test*",
    "activitysim/abm/models/util/test*",
]

[tool.setuptools_scm]
fallback_version = "999"
write_to = "activitysim/_generated_version.py"

[tool.isort]
profile = "black"
skip_gitignore = true
float_to_top = true
default_section = "THIRDPARTY"
known_first_party = "activitysim"
add_imports = "from __future__ import annotations"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --nbmake --disable-warnings --ignore=activitysim/estimation/test"
testpaths = [
    "activitysim/abm/test",
    "activitysim/cli/test",
    "activitysim/core/test",
    "activitysim/abm/models/util/test",
]
filterwarnings = [
    "ignore::tables.NaturalNameWarning",
]
log_cli = true
tb = "native"

[tool.ruff]
line-length = 140
exclude = [
    ".git",
    "__pycache__",
    "docs/_build",
    "docs/conf.py",
    "**/__sharrowcache__",
    "other_resources/",
    "sandbox/"
]

[tool.ruff.lint]
select = ["E", "F", "B", "UP", "TID"]

[tool.ruff.lint.isort]
known-first-party = ["activitysim"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]  # Ignore import violations in all `__init__.py` files
